Check if RSKTAG is missing
Sometimes when getblocktemplate takes long, current works in worklog  don't have the RSKTAG in the coinbase, because delays in threads, for that reason we are checking if the tag exist before submit the solution to RSK node. (eloipool.py)
Also, the RSKTAG is added in coinbaseTxn in case it isn't there. (rootstock.py)